refactor: Refactored src/run.py
